Fix PostgreSQL date arithmetic and schema bugs
- Replace `date - N` integer subtraction with INTERVAL expressions in
  correlations.py (PostgreSQL requires explicit INTERVAL for date math)
- Fix param ordering bug in get_sleep_threshold_vs_performance where
  threshold_milli was bound before since/until, causing type mismatch
- Widen withings_group_id from INTEGER to BIGINT to handle large Withings IDs
- Add COMMANDS.md with common commands for syncs, auth, agent, and DB access

+++ db/postgres_schema.sql
@@ -126,7 +126,7 @@ CREATE TABLE IF NOT EXISTS sleep (
 CREATE TABLE IF NOT EXISTS body_measurements (
     id                SERIAL      PRIMARY KEY,
     user_id           INTEGER     NOT NULL REFERENCES users(id) ON DELETE CASCADE,
-    withings_group_id INTEGER     NOT NULL,
+    withings_group_id BIGINT      NOT NULL,
     measured_at       TIMESTAMPTZ NOT NULL,
     date              DATE        NOT NULL,
     source            TEXT        NOT NULL DEFAULT 'withings',
